+++ .github/workflows/firebase-hosting-merge.yml
@@ -5,7 +5,7 @@ name: Deploy to Firebase Hosting on merge
 'on':
   push:
     branches:
-      - master
+      - ITP-618-web-create-diary-card
 jobs:
   build_and_deploy:
     runs-on: ubuntu-latest
