+++ src/index.js
@@ -7,11 +7,9 @@ import {store} from "./state/store";
 import {Provider } from 'react-redux'
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  <>
     <Provider store={store}>
       <App />
     </Provider>
-  </>
 );
 
 // If you want to start measuring performance in your app, pass a function
